# docs/swagger.yml
swagger: '2.0'
info:
  version: 1.0.0
  title: TomoIssuer APIs
  description: Happy to code TomoIssuer APIs
  license:
    name: Github
    url: 'https://github.com/tomochain/tomoissuer'
tags:
  - name: Config
    description: Get TomoIssuer Application Configuration
  - name: Account
    description: Get account's informations(from TomoScan)
  - name: Token
    description: Get token's information(from TomoScan)
schemes:
  - https
  - http
consumes:
  - application/json
produces:
  - application/json
paths:
  /api/config:
    get:
      tags:
        - Config
      summary: Get TomoIssuer Application Configuration
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '406':
          description: Not Acceptable
        '500':
          description: Server Internal Error
  '/api/account/{address}':
    get:
      tags:
        - Account
      summary: Get account's informations(calling TomoScan API)
      parameters:
        - name: address
          in: path
          type: string
          description: 'Logining address, Owner address'
          required: true
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '406':
          description: Not Acceptable
        '500':
          description: Server Internal Error
  '/api/account/{address}/listTokens':
    get:
      tags:
        - Account
      summary: Get token list following an address(calling TomoScan API)
      parameters:
        - name: address
          in: path
          type: string
          description: 'Logining address, Owner address'
          required: true
        - name: limit
          in: query
          type: number
          description: Limit records in response. Maximum is 200
          minimum: 0
        - name: page
          in: query
          type: number
          description: Page number
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '406':
          description: Not Acceptable
        '500':
          description: Server Internal Error
  /api/token/createToken:
    post:
      tags:
        - Token
      summary: Generate smart contract following inputs
      parameters:
        - name: token
          in: body
          description: Token object
          required: true
          schema:
            type: object
            required:
              - name
              - symbol
              - type
              - totalSupply
              - decimals
            $ref: '#/definitions/token'
      responses:
        '200':
          description: OK. All inputs and token contract source code
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '406':
          description: Not Acceptable
        '500':
          description: Server Internal Error
  '/api/token/{token}':
    get:
      tags:
        - Token
      summary: Get token's informations(calling TomoScan API)
      parameters:
        - name: token
          in: path
          type: string
          description: Token address
          required: true
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '406':
          description: Not Acceptable
        '500':
          description: Server Internal Error
  '/api/token//holders/trc21/{token}':
    get:
      tags:
        - Token
      summary: Get token's holders of trc21 token(calling TomoScan API)
      parameters:
        - name: token
          in: path
          type: string
          description: Token address
          required: true
        - name: limit
          in: query
          type: string
          description: Limit records in response
        - name: page
          in: query
          type: string
          description: Page number
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '406':
          description: Not Acceptable
        '500':
          description: Server Internal Error
  '/api/token//holders/{token}':
    get:
      tags:
        - Token
      summary: Get token's holders of trc20 token(calling TomoScan API)
      parameters:
        - name: token
          in: path
          type: string
          description: Token address
          required: true
        - name: limit
          in: query
          type: string
          description: Limit records in response
        - name: page
          in: query
          type: string
          description: Page number
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '406':
          description: Not Acceptable
        '500':
          description: Server Internal Error
  '/api/token/txs/trc21/{token}':
    get:
      tags:
        - Token
      summary: Get token's transactions of trc21 token(calling TomoScan API)
      parameters:
        - name: token
          in: path
          type: string
          description: Token address
          required: true
        - name: limit
          in: query
          type: string
          description: Limit records in response
        - name: page
          in: query
          type: string
          description: Page number
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '406':
          description: Not Acceptable
        '500':
          description: Server Internal Error
  '/api/token/txs/trc20/{token}':
    get:
      tags:
        - Token
      summary: Get token's transactions of trc20 token(calling TomoScan API)
      parameters:
        - name: token
          in: path
          type: string
          description: Token address
          required: true
        - name: limit
          in: query
          type: string
          description: Limit records in response
        - name: page
          in: query
          type: string
          description: Page number
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '406':
          description: Not Acceptable
        '500':
          description: Server Internal Error
  /api/token/search:
    get:
      tags:
        - Token
      summary: Search token(calling TomoScan API)
      parameters:
        - name: query
          in: query
          type: string
          description: Token name
          required: true
        - name: type
          in: query
          type: string
          description: 'Token type, hardcode: trc21'
        - name: limit
          in: query
          type: string
          description: 'Number of records return, hardcode: 20'
        - name: page
          in: query
          type: string
          description: 'page number, hardcode: 1'
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '406':
          description: Not Acceptable
        '500':
          description: Server Internal Error
definitions:
  token:
    type: object
    example:
      name: Token
      symbol: TOKEN
      decimals: 18
      totalSupply: 100000000
      type: trc21
      minFee: 0.005